Rename files according to LLVM naming convention.
