+++ experimental/iterators/src/Main.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 #include <vector>
 
-#include "operators/column_scan.h"
-#include "utils/tuple.h"
+#include "Operators/ColumnScanOperator.h"
+#include "Utils/Tuple.h"
 
 int main(int /*unused*/, char ** /*unused*/) {
   std::vector<int32_t> numbers = {1, 2, 3, 4};
